refactor!: split out storage lib.rs into another types.rs module
