+++ htsget-search/src/bcf_search.rs
@@ -16,7 +16,8 @@ use tracing::{instrument, trace};
 
 use crate::search::{find_first, BgzfSearch, Search};
 use crate::{Format, Query, Result};
-use htsget_storage::{BytesPosition, Storage, Streamable};
+use htsget_storage::types::BytesPosition;
+use htsget_storage::{Storage, Streamable};
 
 type AsyncReader = bcf::AsyncReader<bgzf::AsyncReader<Streamable>>;
